gemelos_hub/requirements.txt
azure-iot-hub==2.6.1
azure-iot-device==2.14.0




